requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
# mysql-connector-python==9.0.0
pandas
# gspread
# oauth2client
SQLAlchemy
gspread==6.1.4  # or the latest version
oauth2client==4.1.3
pymysql
scikit-learn==1.3.0
